requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0

sqlmodel>=0.0.18
sqlalchemy>=2.0
pymysql

scikit-learn==1.5.2
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
joblib==1.3.2
imblearn==0.0
unidecode==1.3.8

sentence-transformers==2.2.2
huggingface_hub==0.14.1

python-jose==3.3.0
pyjwt==2.8.0
passlib[argon2]==1.7.4
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator>=2.0
